+++ CHANGES.md
@@ -1,3 +1,5 @@
+# UNRELEASED
+
 # 0.16.1 (2017-11-02)
 - [FIXED] Corrected `eslint` to be `devDependencies` instead of `dependencies`.