+++ CHANGES.md
@@ -1,4 +1,5 @@
 # UNRELEASED
+- [FIXED] Abort request retry for dead feed.
 
 # 0.17.0 (2018-03-22)
 - [NEW] Allow a custom HTTP agent to be specified.