+++ CHANGES.md
@@ -1,4 +1,4 @@
-# UNRELEASED
+# 0.18.0 (2018-10-30)
 - [NEW] Support use of changes selector filter.
 - [FIXED] Abort request retry for dead feed.
 - [FIXED] Pipe using decoded request.